+++ packages/svelte/src/server.ts
@@ -1,1 +1,1 @@
-export { default as default } from '@inertiajs/core/server'
+export { default as default } from 'inertiax-core/server'