+++ packages/svelte/src/usePoll.ts
@@ -1,5 +1,5 @@
-import { router, type PollOptions, type ReloadOptions } from '@inertiajs/core'
-import { onDestroy, onMount } from 'svelte'
+import { type PollOptions, type ReloadOptions } from 'inertiax-core'
+import { onDestroy, onMount, getContext } from 'svelte'
 
 export default function usePoll(
   interval: number,
@@ -9,6 +9,7 @@ export default function usePoll(
     autoStart: true,
   },
 ) {
+  const { router } = getContext('frame')
   const { stop, start } = router.poll(interval, requestOptions, {
     ...options,
     autoStart: false,